// apps/admin/src/shared/api/generated/ja-korea/schemas/principleUpdateRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */
import type { PrincipleUpdateItem } from './principleUpdateItem';

export interface PrincipleUpdateRequest {
  /**
     * @minLength 0
     * @maxLength 2000
     */
  introSubText?: string;
  /**
     * @minLength 0
     * @maxLength 20000
     */
  introMainText?: string;
  settingVersion: number;
  /**
     * @minItems 5
     * @maxItems 5
     */
  items: PrincipleUpdateItem[];
}